refactor(lexical): extract EditorPlaceholder into its own file
Per AGENTS.md: one component per file; a file's name must match a symbol it
exports. Moves EditorPlaceholder out of RichTextEditor.tsx.
