+++ packages/lexical-editor/src/components/Editor/RichTextEditor.tsx
@@ -1,6 +1,5 @@
-import React, { Fragment, useEffect, useId, useRef, useState } from "react";
+import React, { Fragment, useId, useRef, useState } from "react";
 import type { Klass, LexicalNode } from "lexical";
-import { $getRoot } from "lexical";
 import { LexicalComposer } from "@lexical/react/LexicalComposer";
 import { AutoFocusPlugin } from "@lexical/react/LexicalAutoFocusPlugin";
 import { ClearEditorPlugin } from "@lexical/react/LexicalClearEditorPlugin";
@@ -14,10 +13,7 @@ import { allNodes } from "@webiny/lexical-nodes";
 import { RichTextEditorProvider } from "~/context/RichTextEditorContext.js";
 import { BlurEventPlugin } from "~/plugins/BlurEventPlugin/BlurEventPlugin.js";
 import type { LexicalValue, ToolbarActionPlugin } from "~/types.js";
-import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import { $isRootTextContentEmpty } from "@lexical/text";
-import { Placeholder } from "~/ui/Placeholder.js";
-import { useRichTextEditor } from "~/hooks/index.js";
+import { EditorPlaceholder } from "~/components/Editor/EditorPlaceholder.js";
 import { SharedHistoryContext, useSharedHistoryContext } from "~/context/SharedHistoryContext.js";
 import {
     LexicalEditorWithConfig,
@@ -51,59 +47,6 @@ export interface RichTextEditorProps {
     configRef?: React.MutableRefObject<InitialEditorConfig | undefined>;
     width?: number | string;
 }
-
-interface EditorPlaceholderProps {
-    text: React.ReactNode;
-    styles?: React.CSSProperties;
-    fallbackClassName?: string;
-}
-
-interface WithStyleId {
-    getStyleId(): string | undefined;
-}
-
-const hasStyleId = (node: LexicalNode | null): node is LexicalNode & WithStyleId => {
-    return !!node && typeof (node as Partial<WithStyleId>).getStyleId === "function";
-};
-
-/**
- * Renders our own placeholder, shown whenever the editor's text content is empty —
- * regardless of the current block type. Lexical's built-in placeholder only shows for an
- * empty paragraph, so it disappears when the block is a heading, quote, etc. The placeholder
- * also picks up the current block's typography (via its style id), falling back to the
- * default paragraph style, so its size/font matches whatever the user is about to type.
- */
-const EditorPlaceholder = ({ text, styles, fallbackClassName }: EditorPlaceholderProps) => {
-    const [editor] = useLexicalComposerContext();
-    const { theme } = useRichTextEditor();
-    const [state, setState] = useState<{ isEmpty: boolean; className?: string }>({
-        isEmpty: true
-    });
-
-    useEffect(() => {
-        const update = () => {
-            editor.getEditorState().read(() => {
-                const isEmpty = $isRootTextContentEmpty(editor.isComposing());
-                const firstBlock = $getRoot().getFirstChild();
-                const styleId = hasStyleId(firstBlock) ? firstBlock.getStyleId() : undefined;
-                const className = styleId ? theme.getTypographyById(styleId)?.className : undefined;
-                setState({ isEmpty, className });
-            });
-        };
-        update();
-        return editor.registerUpdateListener(update);
-    }, [editor, theme]);
-
-    if (!state.isEmpty) {
-        return null;
-    }
-
-    return (
-        <Placeholder styles={styles} className={state.className ?? fallbackClassName}>
-            {text}
-        </Placeholder>
-    );
-};
 
 const BaseRichTextEditor = ({
     onChange,
